+++ src/main/java/org/woehlke/twitterwall/frontend/rest/CountedEntitiesResource.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.woehlke.twitterwall.oodm.entities.parts.AbstractTwitterObject;
+import org.woehlke.twitterwall.oodm.entities.parts.CountedEntities;
 import org.woehlke.twitterwall.scheduled.service.persist.CountedEntitiesService;
 
 /**
@@ -17,7 +17,7 @@ public class CountedEntitiesResource {
 
     @RequestMapping(path="/count",method= RequestMethod.GET)
     public @ResponseBody
-    AbstractTwitterObject.CountedEntities countAll(){
+    CountedEntities countAll(){
         return this.countedEntitiesService.countAll();
     }
