+++ src/main/java/org/woehlke/twitterwall/oodm/service/impl/TaskServiceImpl.java
@@ -3,16 +3,15 @@ package org.woehlke.twitterwall.oodm.service.impl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.woehlke.twitterwall.oodm.entities.Task;
+import org.woehlke.twitterwall.oodm.entities.parts.CountedEntities;
 import org.woehlke.twitterwall.oodm.entities.parts.TaskType;
 import org.woehlke.twitterwall.oodm.dao.TaskDao;
-import org.woehlke.twitterwall.oodm.entities.parts.AbstractTwitterObject;
 import org.woehlke.twitterwall.oodm.repositories.TaskRepository;
 import org.woehlke.twitterwall.oodm.service.TaskService;
 import org.woehlke.twitterwall.scheduled.service.persist.CountedEntitiesService;
@@ -41,6 +40,8 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public Task store(Task domainObject) {
+        return taskRepository.save(domainObject);
+        /*
         try {
             Task taskPersistent = taskDao.findById(domainObject.getId());
             if(domainObject.compareTo(taskPersistent)==0) {
@@ -51,31 +52,36 @@ public class TaskServiceImpl implements TaskService {
         } catch (EmptyResultDataAccessException e) {
             return taskDao.persist(domainObject);
         }
+        */
     }
 
     @Override
     public Task create(Task domainObject) {
-        return taskDao.persist(domainObject);
+        return taskRepository.save(domainObject);
+        //return taskDao.persist(domainObject);
     }
 
     @Override
     public Task update(Task domainObject) {
-        return taskDao.update(domainObject);
+        return taskRepository.save(domainObject);
+        //return taskDao.update(domainObject);
     }
 
     @Override
     public Page<Task> getAll(Pageable pageRequest) {
-        return taskDao.getAll(Task.class,pageRequest);
+        return taskRepository.findAll(pageRequest);
+        //return taskDao.getAll(Task.class,pageRequest);
     }
 
     @Override
     public long count() {
-        return taskDao.count(Task.class);
+        return taskRepository.count();
+        //return taskDao.count(Task.class);
     }
 
     @Override
     public Task create(String msg,TaskType type) {
-        AbstractTwitterObject.CountedEntities countedEntities = this.countedEntitiesService.countAll();
+        CountedEntities countedEntities = this.countedEntitiesService.countAll();
         Task task = new Task(msg,type);
         task.setCountedEntitiesAtStart(countedEntities);
         task.start();
@@ -86,7 +92,7 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public Task done(Task task) {
-        AbstractTwitterObject.CountedEntities countedEntities = this.countedEntitiesService.countAll();
+        CountedEntities countedEntities = this.countedEntitiesService.countAll();
         task.setCountedEntitiesAtFinish(countedEntities);
         task.done();
         task = taskDao.update(task);
@@ -152,6 +158,7 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public Task findById(long id) {
-        return taskDao.findById(id);
+        return taskRepository.findOne(id);
+        //return taskDao.findById(id);
     }
 }
