+++ src/main/java/org/woehlke/twitterwall/scheduled/service/persist/CountedEntitiesService.java
@@ -1,13 +1,13 @@
 package org.woehlke.twitterwall.scheduled.service.persist;
 
-import org.woehlke.twitterwall.oodm.entities.parts.AbstractTwitterObject;
+import org.woehlke.twitterwall.oodm.entities.parts.CountedEntities;
 
 /**
  * Created by tw on 09.07.17.
  */
 public interface CountedEntitiesService {
 
-    AbstractTwitterObject.CountedEntities countAll();
+    CountedEntities countAll();
 
     long countTweets();
